added cto:hasRecommendedTask
The object property cto:hasRecommendedTask has been added as suggestion for CTO. It is a subproperty of cto:isSubjectOfTask. The property could serve as a property for expert- or recommendation-systems, which usually recommend multiple solutions to an end-user. By utilizing this property, the recommendations could be distinguished from planned or executed tasks, which use cto:SubjectOfTask.